Auto stash before merge of "lab_2" and "origin/lab_2"

+++ src/task1-signals.cpp
@@ -5,6 +5,7 @@
 #include <cstdlib>
 #include <ctime>
 #include <vector>
+#include <cerrno>
 
 #include "check.hpp"
 
@@ -21,7 +22,20 @@ volatile sig_atomic_t result_ready = 0;
 volatile sig_atomic_t result = 0;
 
 volatile sig_atomic_t terminate_flag = 0;
+volatile sig_atomic_t peer_dead = 0;
 
+bool is_alive(pid_t pid) {
+  if (kill(pid, 0) == -1) {
+    if (errno == ESRCH) return false;
+  }
+  
+  return true;
+}
+
+void sigchld_handler(int) {
+  peer_dead = 1;
+  terminate_flag = 1;
+}
 
 void guess_handler(int, siginfo_t *info, void *) {
     guess_value = info->si_value.sival_int;
@@ -41,12 +55,12 @@ void setup_handlers() {
     struct sigaction sa_guess{};
     sa_guess.sa_flags = SA_SIGINFO;      
     sa_guess.sa_sigaction = guess_handler;
-    sigemptyset(&sa_guess.sa_mask);      
+    sigfillset(&sa_guess.sa_mask);      
     check(sigaction(SIG_GUESS, &sa_guess, nullptr));
 
     struct sigaction sa_simple{};
     sa_simple.sa_handler = result_handler; 
-    sigemptyset(&sa_simple.sa_mask);
+    sigfillset(&sa_simple.sa_mask);
     sa_simple.sa_flags = 0;          
 
     check(sigaction(SIG_WIN, &sa_simple, nullptr));
@@ -54,15 +68,26 @@ void setup_handlers() {
 
     struct sigaction sa_term{};
     sa_term.sa_handler = term_handler;
-    sigemptyset(&sa_term.sa_mask);
+    sigfillset(&sa_term.sa_mask);
     sa_term.sa_flags = 0;
 
-    check(sigaction(SIGINT, &sa_term, nullptr));
     check(sigaction(SIGTERM, &sa_term, nullptr));
+    
+    struct sigaction sa_chld{};
+    sa_chld.sa_handler = sigchld_handler;
+    sigfillset(&sa_chld.sa_mask);
+    sa_chld.sa_flags = SA_NOCLDWAIT;
+    check(sigaction(SIGCHLD, &sa_chld, nullptr));
 }
 
-void wait_for(volatile sig_atomic_t &flag, sigset_t &wait_mask) {
-    while (!flag && !terminate_flag) {
+void wait_for(volatile sig_atomic_t &flag, sigset_t &wait_mask, pid_t other_pid) {
+    while (!flag && !terminate_flag && !peer_dead) {
+    
+        if (!is_alive(other_pid)) {
+          peer_dead = 1;
+          terminate_flag = 1;
+          break;
+        }
         sigsuspend(&wait_mask);
     }
     flag = 0;
@@ -72,6 +97,12 @@ void send_guess(pid_t pid, int value) {
     union sigval val;
     val.sival_int = value;
     if (sigqueue(pid, SIG_GUESS, val) == -1) {
+    
+        if (errno == ESRCH) {
+            peer_dead = 1;
+            terminate_flag = 1;
+            return;
+        }
         perror("sigqueue");
         exit(1);
     }
@@ -90,11 +121,14 @@ int main(int argc, char *argv[]) {
     sigemptyset(&mask_game);
     sigaddset(&mask_game, SIGRTMIN);
     sigaddset(&mask_game, SIG_WIN); 
-    sigaddset(&mask_game, SIG_LOSE); 
+    sigaddset(&mask_game, SIG_LOSE);
+    sigaddset(&mask_game, SIGCHLD);
 
     check(sigprocmask(SIG_SETMASK, &mask_game, &old_mask));
 
     setup_handlers();
+    
+    pid_t parent_pid = getpid();
 
     pid_t child = fork();
 
@@ -108,7 +142,7 @@ int main(int argc, char *argv[]) {
     bool is_parent = (child != 0);
     const char* role = is_parent ? "Parent" : "Child";
 
-    pid_t other = is_parent ? child : getppid();
+    pid_t other = is_parent ? child : parent_pid;
     
     check(sigprocmask(SIG_SETMASK, &old_mask, nullptr));
     
@@ -132,16 +166,29 @@ int main(int argc, char *argv[]) {
             funlockfile(stdout);
 
             while (!terminate_flag) {
-                wait_for(guess_ready, wait_guess);
+                sleep(1);
+                wait_for(guess_ready, wait_guess, other);
+                if (terminate_flag) break;
 
                 attempts++;
                 
                 flockfile(stdout);
                 cout << "[(" << role << ") PID " << getpid() << "] Получил: " << guess_value << endl << flush;
                 funlockfile(stdout);
+                
+                if (!is_alive(other)) {
+                  peer_dead = 1;
+                  terminate_flag = 1;
+                  break;
+                }
 
                 if (guess_value == secret) {
                     if (kill(other, SIG_WIN) == -1) {
+                      if (errno == ESRCH) {
+                        terminate_flag = 1;
+                        break;
+                      }
+                      
                       perror("kill SIG_WIN");
                       exit(1);
                     }
@@ -153,6 +200,11 @@ int main(int argc, char *argv[]) {
                     break;
                 } else {
                     if (kill(other, SIG_LOSE) == -1) {
+                      if (errno == ESRCH) {
+                        terminate_flag = 1;
+                        break;
+                      }
+                      
                       perror("kill SIG_LOSE");
                       exit(1);
                     }       
@@ -181,10 +233,17 @@ int main(int argc, char *argv[]) {
                 flockfile(stdout);
                 cout << "[(" << role << ") PID " << getpid() << "] Пытаюсь: " << guess << endl << flush;
                 funlockfile(stdout);
-
+                
+                if (!is_alive(other)) {
+                  peer_dead = 1;
+                  terminate_flag = 1;
+                  break;
+                }
+                
                 send_guess(other, guess);
 
-                wait_for(result_ready, wait_result);
+                wait_for(result_ready, wait_result, other);
+                if(terminate_flag) break;
 
                 if (result) {
                     flockfile(stdout);
@@ -198,10 +257,17 @@ int main(int argc, char *argv[]) {
     }
 
     if (is_parent) {
-        kill(child, SIGTERM);
+        if (kill(child, SIGTERM) == -1 && errno != ESRCH) {
+          perror("kill SIGTERM");
+        }
         wait(nullptr);
     }
 
-    cout << "[(" << role << ") PID " << getpid() << "] Завершение" << endl << flush;
+    if (peer_dead) {
+      cout << "[(" << role << ") PID " << getpid() 
+         << "] Партнёр завершился, выходим" << endl << flush;
+    } else {
+      cout << "[(" << role << ") PID " << getpid() << "] Завершение" << endl << flush;
+    }
     return 0;
 }
